Changes in main.py

+++ app/main.py
@@ -1,25 +1,26 @@
 import json
-
+import os
 from app import address, tools, mail
 import time
+from tools import json_load, HOME
 
 
 def search_books():
-    path = "searching_books.json"
+    path = HOME / "searching_books.json"
     while True:
-        with open(path) as file:
-            book_list = json.load(file)
-        for title, email in book_list.items():
-            # convert it into SINGLE url.
-            url = address.get_url(title)
-            # checking status for book
-            availability = address.check_for_book_status(url)
-            # if it is available, send mail and delete from searching_titles.
-            if availability:
-                mail.send_mail(title, email)
-                book_list.pop(title)
-        with open("searching_books.json" "w") as file:
-            json.dump(book_list, file)
+        if os.path.isfile(path):
+            book_list = json_load(path)
+            titles_to_remove = []
+            for title, email in book_list.items():
+                # convert it into SINGLE url.
+                url = address.get_url(title)
+                # checking status for book
+                availability = address.check_for_book_status(url)
+                # if it is available, send mail and delete from searching_titles.
+                if availability:
+                    mail.send_mail(title, email)
+                    titles_to_remove.append(title)
+            tools.delete_from_searching_book_file(titles_to_remove, path)
         print("Sleeping...")
         time.sleep(60 * 60 * 12)
